Move date filter component to shared location
