+++ src/js/i18n/en-US.yml
@@ -448,6 +448,25 @@ careOptsFrontend:
           postBtn: Post
           saveBtn: Save
       components:
+        dateFilterComponent:
+          dateFilterViews:
+            actionViews:
+              month: This Month
+              today: Today
+              yesterday: Yesterday
+            datepickerLayoutView:
+              added: Added
+              due: Due
+              updated: Updated
+            dateTypes:
+              created_at: 'Added:'
+              due_date: 'Due:'
+              updated_at: 'Updated:'
+            defaultDateView:
+              thisMonth: This Month
+            relativeDateView:
+              today: Today
+              yesterday: Yesterday
         dueComponent:
           defaultText: Select Date...
         durationComponent:
@@ -630,25 +649,6 @@ careOptsFrontend:
           bodyText: Are you sure you want to delete the selected Flows? This cannot be undone.
           headingText: Delete Flows?
           submitText: Delete
-      filters:
-        dateFilterViews:
-          actionViews:
-            month: This Month
-            today: Today
-            yesterday: Yesterday
-          datepickerLayoutView:
-            added: Added
-            due: Due
-            updated: Updated
-          dateTypes:
-            created_at: 'Added:'
-            due_date: 'Due:'
-            updated_at: 'Updated:'
-          defaultDateView:
-            thisMonth: This Month
-          relativeDateView:
-            today: Today
-            yesterday: Yesterday
       filtersApp:
         allGroupsName: All Groups
         ownerFilterHeadingText: Filter by Owner
