+++ src/js/views/patients/shared/components/date-filter/date-filter.scss
